fix: protect delete

+++ pages/api/posts/[id]/delete.ts
@@ -1,11 +1,24 @@
 import prisma from "lib/prisma";
 import { NextApiHandler } from "next";
+import { getSession } from "next-auth/client";
 
 const deletePost: NextApiHandler = async (req, res) => {
+    const session = await getSession({req})
     const {
         query: {id},
         method
     } = req;
+    const post = await prisma.post.findUnique({
+        include: {author: true},
+        where: {id: Number(id)}
+    })
+    if (!session || session?.user?.email !== post?.author.email) {
+        res.status(401).end("You don't have permission to delete this post");
+        return;
+        // throw new Error (
+        //     `You don't have permission to delete this post`
+        // )
+    }
     if (method === "DELETE") {
         const post = await prisma.post.delete({
             where: {id: Number(id)}
